pyproject.toml
[project]
name = "job-scout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "patchright>=1.57.2",
    "playwright>=1.57.0",
    "pytelegrambotapi>=4.29.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
